separate Logic traits to independent scala files.
