+++ src/main/scala/com/github/everpeace/k8s/throttler/controller/ThrottleController.scala
@@ -28,12 +28,9 @@ import com.github.everpeace.k8s.throttler.crd.v1alpha1
 import com.github.everpeace.k8s.throttler.crd.v1alpha1.Implicits._
 import com.github.everpeace.k8s.throttler.metrics.{
   ClusterThrottleControllerMetrics,
-  MetricsBase,
   ThrottleControllerMetrics
 }
-import kamon.Kamon
 import play.api.libs.json.Format
-import skuber.Resource.ResourceList
 import skuber.ResourceSpecification.Subresources
 import skuber.api.client.EventType
 import skuber.json.format._
@@ -560,83 +557,5 @@ object ThrottleController {
     implicit val _k8s    = k8s
     implicit val _config = config
     Props(new ThrottleController())
-  }
-}
-
-trait ClusterThrottleControllerLogic {
-
-  def isClusterThrottleActiveFor(pod: Pod, clthrottle: v1alpha1.ClusterThrottle): Boolean =
-    clthrottle.isActiveFor(pod)
-  def isClusterThrottleInsufficientFor(pod: Pod, clthrottle: v1alpha1.ClusterThrottle): Boolean =
-    clthrottle.isInsufficientFor(pod)
-  def calcNextClusterThrottleStatuses(
-      targetClusterThrottles: Set[v1alpha1.ClusterThrottle],
-      podsInAllNamespaces: Set[Pod]
-    ): List[(ObjectKey, v1alpha1.ClusterThrottle.Status)] = {
-
-    for {
-      clthrottle <- targetClusterThrottles.toList
-
-      matchedPods = podsInAllNamespaces.filter(p =>
-        clthrottle.spec.selector.matches(p.metadata.labels))
-      running = matchedPods.filter(p => p.status.exists(_.phase.exists(_ == Pod.Phase.Running)))
-      usedResource = v1alpha1.ResourceAmount(
-        podsCount = clthrottle.spec.threshold.podsCount.map(_ => running.size),
-        resourceRequests = {
-          val actual =
-            running.toList.map(_.totalRequests).foldLeft(Map.empty: ResourceList)(_ add _)
-          val ret = for {
-            (r, _) <- clthrottle.spec.threshold.resourceRequests if actual.contains(r)
-          } yield r -> actual(r)
-          ret
-        }
-      )
-      nextStatus = clthrottle.spec.statusFor(usedResource)
-
-      toUpdate <- if (clthrottle.status != Option(nextStatus)) {
-                   List(clthrottle.key -> nextStatus)
-                 } else {
-                   List.empty
-                 }
-    } yield toUpdate
-  }
-}
-
-trait ThrottleControllerLogic {
-  def isThrottleActiveFor(pod: Pod, throttle: v1alpha1.Throttle): Boolean =
-    throttle.isThrottleActiveFor(pod)
-
-  def isThrottleInsufficientFor(pod: Pod, throttle: v1alpha1.Throttle): Boolean =
-    throttle.isThrottleInsufficientFor(pod)
-
-  def calcNextThrottleStatuses(
-      targetThrottles: Set[v1alpha1.Throttle],
-      podsInNs: Set[Pod]
-    ): List[(ObjectKey, v1alpha1.Throttle.Status)] = {
-
-    for {
-      throttle <- targetThrottles.toList
-
-      matchedPods = podsInNs.filter(p => throttle.spec.selector.matches(p.metadata.labels))
-      running     = matchedPods.filter(p => p.status.exists(_.phase.exists(_ == Pod.Phase.Running)))
-      usedResource = v1alpha1.ResourceAmount(
-        podsCount = throttle.spec.threshold.podsCount.map(_ => running.size),
-        resourceRequests = {
-          val actual =
-            running.toList.map(_.totalRequests).foldLeft(Map.empty: ResourceList)(_ add _)
-          val ret = for {
-            (r, _) <- throttle.spec.threshold.resourceRequests if actual.contains(r)
-          } yield r -> actual(r)
-          ret
-        }
-      )
-      nextStatus = throttle.spec.statusFor(usedResource)
-
-      toUpdate <- if (throttle.status != Option(nextStatus)) {
-                   List(throttle.key -> nextStatus)
-                 } else {
-                   List.empty
-                 }
-    } yield toUpdate
   }
 }
